+++ src/rust/src/backend/mod.rs
@@ -21,6 +21,7 @@ pub(crate) mod hmac;
 pub(crate) mod hpke;
 pub(crate) mod kdf;
 pub(crate) mod keys;
+pub(crate) mod keywrap;
 #[cfg(any(
     CRYPTOGRAPHY_IS_BORINGSSL,
     CRYPTOGRAPHY_IS_AWSLC,